Type: Move smartdict in separate class.
