+++ tests/test_type.py
@@ -50,23 +50,6 @@ class TestType(unittest.TestCase):
         # FIXME: This shouldn't be useful anymore.
         TestType.parser.logger.clear()
 
-    def test_smartdict(self):
-        sd = type.Table.smartdict()
-        t = ast.User("foo")
-        t.lineno, t.lexpos = 1, 1
-        sd[t] = "foo"
-        sd[t].should.be.equal("foo")
-
-        tt = ast.User("foo")
-        tt.lineno, tt.lexpos = 0, 0
-        ttt = sd.getKey(tt)
-        ttt.lineno.should.be.equal(1)
-        ttt.lexpos.should.be.equal(1)
-
-        del sd[t]
-        sd.get(t).should.be(None)
-        sd.getKey(t).should.be(None)
-
     def test_type_process(self):
         right_testcases = (
             "type color = Red | Green | Blue",
